init: hello world, flux-mcp!
This first commit has the core structure, with a registry and
support for two server types - fastmcp (http default but can
also serve sse (deprecated) and stdio) along with more vanilla
mcp. I have added the first two tools, validate and count, and
an example to run the server and tests for it. I will add the count
tests next, and then go through the list of functions we want for
flux one at a time (and add to git when the testing is added and
done). After that I will set this all up to run and test in CI
and we should be good!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>